// frontend/fingertips-frontend/generated-sources/ft-api-client/models/IndicatorPolarity.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Fingertips API
 * An API to query public health indicator data.
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: ProfileFeedback@dhsc.gov.uk
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const IndicatorPolarity = {
    Unknown: 'Unknown',
    NoJudgement: 'NoJudgement',
    LowIsGood: 'LowIsGood',
    HighIsGood: 'HighIsGood'
} as const;
export type IndicatorPolarity = typeof IndicatorPolarity[keyof typeof IndicatorPolarity];


export function instanceOfIndicatorPolarity(value: any): boolean {
    for (const key in IndicatorPolarity) {
        if (Object.prototype.hasOwnProperty.call(IndicatorPolarity, key)) {
            if (IndicatorPolarity[key as keyof typeof IndicatorPolarity] === value) {
                return true;
            }
        }
    }
    return false;
}

export function IndicatorPolarityFromJSON(json: any): IndicatorPolarity {
    return IndicatorPolarityFromJSONTyped(json, false);
}

export function IndicatorPolarityFromJSONTyped(json: any, ignoreDiscriminator: boolean): IndicatorPolarity {
    return json as IndicatorPolarity;
}

export function IndicatorPolarityToJSON(value?: IndicatorPolarity | null): any {
    return value as any;
}

export function IndicatorPolarityToJSONTyped(value: any, ignoreDiscriminator: boolean): IndicatorPolarity {
    return value as IndicatorPolarity;
}